Split 'essenceStream' folder into 'decoder' and 'encoder' folders
Also add SWIG interface for each directory.
