+++ src/AvTranscoder/decoder/AudioGenerator.hpp
@@ -1,5 +1,5 @@
-#ifndef _AV_TRANSCODER_ESSENCE_STREAM_GENERATOR_AUDIO_HPP_
-#define _AV_TRANSCODER_ESSENCE_STREAM_GENERATOR_AUDIO_HPP_
+#ifndef _AV_TRANSCODER_DECODER_AUDIO_GENERATOR_HPP_
+#define _AV_TRANSCODER_DECODER_AUDIO_GENERATOR_HPP_
 
 #include "IDecoder.hpp"
 #include <AvTranscoder/codec/AudioCodec.hpp>
