+++ src/AvTranscoder/decoder/VideoDecoder.hpp
@@ -1,5 +1,5 @@
-#ifndef _AV_TRANSCODER_ESSENCE_STREAM_AV_INPUT_VIDEO_HPP_
-#define _AV_TRANSCODER_ESSENCE_STREAM_AV_INPUT_VIDEO_HPP_
+#ifndef _AV_TRANSCODER_DECODER_VIDEO_DECODER_HPP_
+#define _AV_TRANSCODER_DECODER_VIDEO_DECODER_HPP_
 
 #include "IDecoder.hpp"
 #include <AvTranscoder/ProfileLoader.hpp>
