+++ src/AvTranscoder/decoder/VideoGenerator.hpp
@@ -1,5 +1,5 @@
-#ifndef _AV_TRANSCODER_ESSENCE_STREAM_GENERATOR_VIDEO_HPP_
-#define _AV_TRANSCODER_ESSENCE_STREAM_GENERATOR_VIDEO_HPP_
+#ifndef _AV_TRANSCODER_DECODER_VIDEO_GENERATOR_HPP_
+#define _AV_TRANSCODER_DECODER_VIDEO_GENERATOR_HPP_
 
 #include "IDecoder.hpp"
 #include <AvTranscoder/codec/VideoCodec.hpp>
