+++ src/AvTranscoder/decoder/decoder.i
@@ -1,0 +1,13 @@
+%{
+#include <AvTranscoder/decoder/IDecoder.hpp>
+#include <AvTranscoder/decoder/AudioDecoder.hpp>
+#include <AvTranscoder/decoder/VideoDecoder.hpp>
+#include <AvTranscoder/decoder/VideoGenerator.hpp>
+#include <AvTranscoder/decoder/AudioGenerator.hpp>
+%}
+
+%include <AvTranscoder/decoder/IDecoder.hpp>
+%include <AvTranscoder/decoder/AudioDecoder.hpp>
+%include <AvTranscoder/decoder/VideoDecoder.hpp>
+%include <AvTranscoder/decoder/VideoGenerator.hpp>
+%include <AvTranscoder/decoder/AudioGenerator.hpp>
