+++ src/AvTranscoder/encoder/AudioEncoder.hpp
@@ -1,5 +1,5 @@
-#ifndef _AV_TRANSCODER_ESSENCE_STREAM_AV_OUTPUT_AUDIO_HPP_
-#define _AV_TRANSCODER_ESSENCE_STREAM_AV_OUTPUT_AUDIO_HPP_
+#ifndef _AV_TRANSCODER_ENCODER_AUDIO_ENCODER_HPP_
+#define _AV_TRANSCODER_ENCODER_AUDIO_ENCODER_HPP_
 
 #include "IEncoder.hpp"
 #include <AvTranscoder/codec/AudioCodec.hpp>
