+++ src/AvTranscoder/encoder/VideoEncoder.hpp
@@ -1,5 +1,5 @@
-#ifndef _AV_TRANSCODER_ESSENCE_STREAM_AV_OUTPUT_VIDEO_HPP_
-#define _AV_TRANSCODER_ESSENCE_STREAM_AV_OUTPUT_VIDEO_HPP_
+#ifndef _AV_TRANSCODER_ENCODER_VIDEO_ENCODER_HPP_
+#define _AV_TRANSCODER_ENCODER_VIDEO_ENCODER_HPP_
 
 #include "IEncoder.hpp"
 #include <AvTranscoder/codec/VideoCodec.hpp>
