+++ src/AvTranscoder/encoder/encoder.i
@@ -1,0 +1,9 @@
+%{
+#include <AvTranscoder/encoder/IEncoder.hpp>
+#include <AvTranscoder/encoder/AudioEncoder.hpp>
+#include <AvTranscoder/encoder/VideoEncoder.hpp>
+%}
+
+%include <AvTranscoder/encoder/IEncoder.hpp>
+%include <AvTranscoder/encoder/AudioEncoder.hpp>
+%include <AvTranscoder/encoder/VideoEncoder.hpp>
